chore: format gef files same as broxml
